requirements.txt
Django==3.2
djangorestframework_jwt==1.11.0
djangorestframework_simplejwt==4.6.0
djangorestframework==3.12.4
guincorn